source/electrical_heating/tests/container_tests/requirements.txt
# This is a pip 'requirements.txt' file.
# See https://pip.pypa.io/en/stable/reference/requirements-file-format/

# Pytest is installed by default in the container, so it does not need to be listed here.

opengeh-testcommon @ git+https://github.com/Energinet-DataHub/opengeh-python-packages@3.3.0#subdirectory=source/testcommon
pydantic-settings >=2.7.1